refactor(identify_opponent): split MTGOpponentDeckSpy into mixin package
Move widgets/frames/identify_opponent.py into a package: frame.py keeps
the sys.path bootstrap, the _LoadArchetypeDialog, the __init__/_build_ui
plus calculator-panel builder, and the main() CLI entry; properties.py
holds _t, _is_widget_ok, get_latest_deck, and the LEGACY_* path
constants (still monkeypatched by tests via the package namespace);
handlers.py owns every _on_* callback plus the radar worker, the
BackgroundWorker polling pipeline, persistence, and placement helpers.

Handlers looks LEGACY_* up via the package module (_pkg.LEGACY_*) so
tests/ui/conftest.py's monkeypatches on widgets.frames.identify_opponent
reach the runtime consumers. pyproject.toml's ruff E402 ignore is
repointed to the new frame.py path.
