+++ pyproject.toml
@@ -68,7 +68,7 @@ ignore = [
 "test_*.py" = ["C901"]
 "**/tests/**" = ["C901"]
 # Widget entry points insert project root into sys.path before other imports
-"widgets/frames/identify_opponent.py" = ["E402"]
+"widgets/frames/identify_opponent/frame.py" = ["E402"]
 "widgets/frames/match_history/frame.py" = ["E402"]
 "widgets/frames/radar/frame.py" = ["E402"]
 "widgets/frames/timer_alert/frame.py" = ["E402"]
